windows/requirements.txt
Pillow>=12.0.0
pystray>=0.19.5
requests>=2.32.0
